setQuiet -> cli.set_quiet
